+++ pack.sh
@@ -2,7 +2,7 @@
 source ./build.sh
 dotnet pack ./src/aardpack/aardpack.fsproj -c Release -o bin/pack
 
-dotnet publish ./src/Aardvark.Build.Tool/Aardvark.Build.Tool.fsproj -c Release -o bin/publish/standalone-tool
+dotnet publish ./src/Aardvark.Build/Aardvark.Build.fsproj -c Release -o bin/publish/standalone-tool
 
 version=`./bin/Release/net8.0/aardpack.exe --parse-only`
 dotnet paket pack --template ./src/Aardvark.Build/paket.template --version $version bin/pack