+++ src/Tests/TestApp/RELEASE_NOTES.md
@@ -3,4 +3,5 @@
 
 ### 9.9.9
 - Test
+- Include ; some ; semicolons
 - Foo